FIX GroupXData TrueParams now use input data types, rather than casting all int32 arrays to float64

+++ bnpy/data/GroupXData.py
@@ -121,7 +121,7 @@ class GroupXData(XData):
         if TrueParams is not None:
             self.TrueParams = dict()
             for key, arr in TrueParams.items():
-                self.TrueParams[key] = toCArray(arr)
+                self.TrueParams[key] = toCArray(arr, dtype=arr.dtype)
 
         if TrueZ is not None:
             if not hasattr(self, 'TrueParams'):
